requirements/requirements-all.txt
peppy>=0.32
fastapi>=0.72.0
eido>=0.1.5
logmuse>=0.2.7
jinja2>=3.1.2
yacman>=0.8.4
python-multipart>=0.0.5
tqdm
uvicorn
python-dotenv
pepagent @ git+https://github.com/pepkit/pephub_db.git#egg=pepagent
peppy @ git+https://github.com/pepkit/peppy.git@dev_alex#egg=peppy